fix: CI 编译错误修复——UnsafeCell Sync/camelCase→snake_case/平台类型兼容
CI 工作流 #29197496204 报 41 个编译错误，6 个根因全部修复：

1. E0277: UnsafeCell 非 Sync 不能用作 static——添加 unsafe impl Sync
   (hooks.rs OrigFuncs, jni_env.rs JniOffsetsCell, proxy_gen.rs)
2. E0599: gen_jni_hooks.py 生成 camelCase 方法名，HookContext 用 snake_case
   ——加 camel_to_snake() 转换
3. E0507: for entry in JNI_METHOD_TABLE 从 static move——改为 & 引用遍历
4. E0596: addr_of_mut! 参数非 mut——加 rust_mut_params() 生成带 mut 签名
5. E0308/E0277: fd.rs i8 vs u8 比较 + ipc.rs socklen_t 平台类型差异
6. 举一反三：全库审查同类模式，确认 0 处残留

+++ native/src/zygisk_inject/src/fd.rs
@@ -72,8 +72,8 @@ pub fn record_open_fds(fds: &mut FdSet) {
         let mut valid = true;
         for &c in name.iter() {
             if c == 0 { break; }
-            if c >= b'0' && c <= b'9' {
-                fd_val = fd_val * 10 + (c - b'0') as i32;
+            if (c as u8) >= b'0' && (c as u8) <= b'9' {
+                fd_val = fd_val * 10 + (c as u8 - b'0') as i32;
             } else { valid = false; break; }
         }
         if valid && fd_val >= 0 && (fd_val as usize) < MAX_FD_SIZE {
